requirements.txt
langchain
langchain_core
langchain_community
fastapi
uvicorn
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
lxml
langchain-openai
langchain_chroma
beautifulsoup4
faiss-cpu
streamlit
langchain_groq
langchain_core
langserve
sse_starlette